cleanup: remove unused code
