+++ devimint/Cargo.toml
@@ -22,7 +22,6 @@ chrono = { workspace = true }
 clap = { workspace = true }
 esplora-client = { workspace = true }
 fedimint-api-client = { workspace = true }
-fedimint-bitcoind = { workspace = true }
 fedimint-client = { workspace = true }
 fedimint-client-module = { workspace = true }
 fedimint-core = { workspace = true }
