+++ fedimint-bitcoind/src/lib.rs
@@ -10,37 +10,13 @@ use std::env;
 use std::fmt::Debug;
 use std::sync::Arc;
 
-use anyhow::Result;
-use bitcoin::{Block, BlockHash, ScriptBuf, Transaction, Txid};
+use anyhow::{Result, format_err};
+use bitcoin::{ScriptBuf, Transaction, Txid};
+use esplora_client::{AsyncClient, Builder};
 use fedimint_core::envs::{BitcoinRpcConfig, FM_FORCE_BITCOIN_RPC_URL_ENV};
 use fedimint_core::txoproof::TxOutProof;
 use fedimint_core::util::SafeUrl;
-use fedimint_core::{Feerate, apply, async_trait_maybe_send};
-
-#[cfg(feature = "bitcoincore-rpc")]
-pub mod bitcoincore;
-#[cfg(feature = "esplora-client")]
-mod esplora;
-mod feerate_source;
-
-#[cfg(feature = "fedimint-server")]
-pub mod shared;
-
-use crate::esplora::EsploraClient;
-
-// <https://blockstream.info/api/block-height/0>
-const MAINNET_GENESIS_BLOCK_HASH: &str =
-    "000000000019d6689c085ae165831e934ff763ae46a2a6c172b3f1b60a8ce26f";
-// <https://blockstream.info/testnet/api/block-height/0>
-const TESTNET_GENESIS_BLOCK_HASH: &str =
-    "000000000933ea01ad0ee984209779baaec3ced90fa3f408719526f8d77f4943";
-// <https://mempool.space/signet/api/block-height/0>
-const SIGNET_GENESIS_BLOCK_HASH: &str =
-    "00000008819873e925422c1ff0f99f7cc9bbb232af63a077a480a3633bee1ef6";
-// See <https://bitcoin.stackexchange.com/questions/122778/is-the-regtest-genesis-hash-always-the-same-or-not>
-// <https://github.com/bitcoin/bitcoin/blob/d82283950f5ff3b2116e705f931c6e89e5fdd0be/src/kernel/chainparams.cpp#L478>
-const REGTEST_GENESIS_BLOCK_HASH: &str =
-    "0f9188f13cb7b2c71f2a335e3a4fc328bf5beb436012afca590b1a11466e2206";
+use fedimint_core::{apply, async_trait_maybe_send};
 
 /// Create a bitcoin RPC of a given kind
 pub fn create_bitcoind(config: &BitcoinRpcConfig) -> Result<DynBitcoindRpc> {
@@ -60,61 +36,10 @@ pub type DynBitcoindRpc = Arc<dyn IBitcoindRpc + Send + Sync>;
 /// Functions may panic if the bitcoind node is not reachable.
 #[apply(async_trait_maybe_send!)]
 pub trait IBitcoindRpc: Debug + Send + Sync + 'static {
-    /// Returns the Bitcoin network the node is connected to
-    async fn get_network(&self) -> Result<bitcoin::Network>;
-
-    /// Returns the current block count
-    async fn get_block_count(&self) -> Result<u64>;
-
-    /// Returns the block hash at a given height
-    ///
-    /// # Panics
-    /// If the node does not know a block for that height. Make sure to only
-    /// query blocks of a height less to the one returned by
-    /// `Self::get_block_count`.
-    ///
-    /// While there is a corner case that the blockchain shrinks between these
-    /// two calls (through on average heavier blocks on a fork) this is
-    /// prevented by only querying hashes for blocks tailing the chain tip
-    /// by a certain number of blocks.
-    async fn get_block_hash(&self, height: u64) -> Result<BlockHash>;
-
-    async fn get_block(&self, block_hash: &BlockHash) -> Result<Block>;
-
-    /// Estimates the fee rate for a given confirmation target. Make sure that
-    /// all federation members use the same algorithm to avoid widely
-    /// diverging results. If the node is not ready yet to return a fee rate
-    /// estimation this function returns `None`.
-    async fn get_fee_rate(&self, confirmation_target: u16) -> Result<Option<Feerate>>;
-
-    /// Submits a transaction to the Bitcoin network
-    ///
-    /// This operation does not return anything as it never OK to consider its
-    /// success as final anyway. The caller should be retrying
-    /// broadcast periodically until it confirms the transaction was actually
-    /// via other means or decides that is no longer relevant.
-    ///
-    /// Also - most backends considers brodcasting a tx that is already included
-    /// in the blockchain as an error, which breaks idempotency and requires
-    /// brittle workarounds just to reliably ignore... just to retry on the
-    /// higher level anyway.
-    ///
-    /// Implementations of this error should log errors for debugging purposes
-    /// when it makes sense.
-    async fn submit_transaction(&self, transaction: Transaction);
-
     /// If a transaction is included in a block, returns the block height.
     /// Note: calling this method with bitcoind as a backend must first call
     /// `watch_script_history` or run bitcoind with txindex enabled.
     async fn get_tx_block_height(&self, txid: &Txid) -> Result<Option<u64>>;
-
-    /// Check if a transaction is included in a block
-    async fn is_tx_in_block(
-        &self,
-        txid: &Txid,
-        block_hash: &BlockHash,
-        block_height: u64,
-    ) -> Result<bool>;
 
     /// Watches for a script and returns any transactions associated with it
     ///
@@ -133,17 +58,69 @@ pub trait IBitcoindRpc: Debug + Send + Sync + 'static {
     /// Returns a proof that a tx is included in the bitcoin blockchain
     async fn get_txout_proof(&self, txid: Txid) -> Result<TxOutProof>;
 
-    /// Returns the node's estimated chain sync percentage as a float between
-    /// 0.0 and 1.0, or `None` if the node doesn't support this feature.
-    async fn get_sync_percentage(&self) -> Result<Option<f64>>;
-
-    /// Returns the Bitcoin RPC config
-    fn get_bitcoin_rpc_config(&self) -> BitcoinRpcConfig;
-
     fn into_dyn(self) -> DynBitcoindRpc
     where
         Self: Sized,
     {
         Arc::new(self)
+    }
+}
+
+#[derive(Debug)]
+pub struct EsploraClient {
+    client: AsyncClient,
+}
+
+impl EsploraClient {
+    pub fn new(url: &SafeUrl) -> anyhow::Result<Self> {
+        Ok(Self {
+            // URL needs to have any trailing path including '/' removed
+            client: Builder::new(url.as_str().trim_end_matches('/')).build_async()?,
+        })
+    }
+}
+
+#[apply(async_trait_maybe_send!)]
+impl IBitcoindRpc for EsploraClient {
+    async fn get_tx_block_height(&self, txid: &Txid) -> anyhow::Result<Option<u64>> {
+        Ok(self
+            .client
+            .get_tx_status(txid)
+            .await?
+            .block_height
+            .map(u64::from))
+    }
+
+    async fn watch_script_history(&self, _: &ScriptBuf) -> anyhow::Result<()> {
+        // no watching needed, has all the history already
+        Ok(())
+    }
+
+    async fn get_script_history(
+        &self,
+        script: &ScriptBuf,
+    ) -> anyhow::Result<Vec<bitcoin::Transaction>> {
+        let transactions = self
+            .client
+            .scripthash_txs(script, None)
+            .await?
+            .into_iter()
+            .map(|tx| tx.to_tx())
+            .collect::<Vec<_>>();
+
+        Ok(transactions)
+    }
+
+    async fn get_txout_proof(&self, txid: Txid) -> anyhow::Result<TxOutProof> {
+        let proof = self
+            .client
+            .get_merkle_block(&txid)
+            .await?
+            .ok_or(format_err!("No merkle proof found"))?;
+
+        Ok(TxOutProof {
+            block_header: proof.header,
+            merkle_proof: proof.txn,
+        })
     }
 }
