+++ modules/fedimint-lnv2-server/Cargo.toml
@@ -15,7 +15,6 @@ anyhow = { workspace = true }
 async-trait = { workspace = true }
 bls12_381 = { workspace = true }
 erased-serde = { workspace = true }
-fedimint-bitcoind = { workspace = true }
 fedimint-core = { workspace = true }
 fedimint-lnv2-common = { workspace = true }
 fedimint-logging = { workspace = true }
