+++ modules/fedimint-wallet-server/Cargo.toml
@@ -21,7 +21,6 @@ async-trait = { workspace = true }
 bitcoin = { workspace = true }
 erased-serde = { workspace = true }
 fedimint-api-client = { workspace = true }
-fedimint-bitcoind = { workspace = true }
 fedimint-core = { workspace = true }
 fedimint-logging = { workspace = true }
 fedimint-metrics = { workspace = true }
